Coerce to a tsibble object

Coerce to a tsibble object

# S3 method for demogdata
as_tsibble(x, ..., validate = TRUE)

Arguments

x

Objects to be coerced to a tsibble (`tbl_ts`).

...

Other arguments passed on to individual methods.

validate

`TRUE` suggests to verify that each key or each combination of key variables leads to unique time indices (i.e. a valid tsibble). If you are sure that it's a valid input, specify `FALSE` to skip the checks.

Value

A tsibble object.

See also

Examples

# coerce demogdata object to tsibble ----
as_tsibble(demography::fr.mort)
#> # A tsibble: 63,603 x 7 [1Y]
#> # Key:       AgeGroup, Age, Group [333]
#>     Year AgeGroup   Age Group  Mortality Exposure Deaths
#>    <int> <chr>    <int> <chr>      <dbl>    <dbl>  <dbl>
#>  1  1816 0            0 female   0.187    408224. 76332.
#>  2  1816 1            1 female   0.0467   382452. 17861.
#>  3  1816 2            2 female   0.0339   351454. 11924.
#>  4  1816 3            3 female   0.0229   337733.  7738.
#>  5  1816 4            4 female   0.0160   331576.  5304.
#>  6  1816 5            5 female   0.0138   313554.  4338.
#>  7  1816 6            6 female   0.0121   306853.  3714.
#>  8  1816 7            7 female   0.0104   301860.  3148.
#>  9  1816 8            8 female   0.00891  296639.  2642.
#> 10  1816 9            9 female   0.00760  289057.  2195.
#> # ℹ 63,593 more rows