Compute life expectancy from mortality rates

Returns remaining life expectancy at a given age (0 by default).

life_expectancy(.data, from_age = 0, age, sex, mortality)

Arguments

.data

A tsibble including an age variable and a variable containing mortality rates.

from_age

Age at which life expectancy to be calculated. Either a scalar or a vector of ages.

age

Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` or `Age_group` will be used (not case sensitive).

sex

Optional variable in `.data` containing sex information. If omitted, the variable with name `Sex` or `Group` will be used (not case sensitive).

mortality

Variable in `.data` containing Mortality rates (mx). If omitted, the variable with name `mx`, `Mortality` or `Rate` will be used (not case sensitive).

Value

A tsibble object with life expectancy in column `ex`.

References

Chiang CL. (1984) The life table and its applications. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) Applied mathematical demography, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) Demography: measuring and modeling population processes. Blackwell

See also

Examples

# Compute Australia life expectancy for females over time
aus_mortality |>
  dplyr::filter(Code == "AUS", Sex == "female") |>
  life_expectancy()
#> # A tsibble: 103 x 6 [1Y]
#> # Key:       Age, Sex, State, Code [1]
#>     Year Sex    State     Code     ex   Age
#>    <int> <chr>  <chr>     <chr> <dbl> <int>
#>  1  1901 female Australia AUS    56.3     0
#>  2  1902 female Australia AUS    55.8     0
#>  3  1903 female Australia AUS    56.4     0
#>  4  1904 female Australia AUS    58.7     0
#>  5  1905 female Australia AUS    59.8     0
#>  6  1906 female Australia AUS    59.5     0
#>  7  1907 female Australia AUS    59.6     0
#>  8  1908 female Australia AUS    59.8     0
#>  9  1909 female Australia AUS    61.4     0
#> 10  1910 female Australia AUS    61.2     0
#> # ℹ 93 more rows