Compute period lifetables from mortality rates

All available years and ages are included in the tables. qx = mx/(1 + ((1-ax) * mx)) as per Chiang (1984). Warning: the code has only been tested for data based on single-year age groups.

life_table(.data, age, sex, mortality)

Arguments

.data

A tsibble including an age variable and a variable containing mortality rates.

age

Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` or `Age_group` will be used (not case sensitive).

sex

Optional variable in `.data` containing sex information. If omitted, the variable with name `Sex` or `Group` will be used (not case sensitive).

mortality

Variable in `.data` containing Mortality rates (mx). If omitted, the variable with name `mx`, `Mortality` or `Rate` will be used (not case sensitive).

Value

A tsibble object containing the index, keys, and the new life table variables `mx`, `qx`, `lx`, `dx`, `Lx`, `Tx` and `ex`.

References

Chiang CL. (1984) The life table and its applications. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) Applied mathematical demography, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) Demography: measuring and modeling population processes. Blackwell

Examples

# Compute Australia life table for females in 2003
aus_mortality |>
  dplyr::filter(Code == "AUS", Sex == "female", Year == 2003) |>
  life_table()
#> # A tsibble: 101 x 12 [?]
#> # Key:       Age, Sex, State, Code [101]
#>     Year Sex   State Code       mx      qx    lx      dx    Lx    Tx    ex   Age
#>    <int> <chr> <chr> <chr>   <dbl>   <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <int>
#>  1  2003 fema… Aust… AUS   4.33e-3 4.32e-3 1     4.32e-3 0.996  83.2  83.2     0
#>  2  2003 fema… Aust… AUS   3.66e-4 3.66e-4 0.996 3.64e-4 0.996  82.2  82.5     1
#>  3  2003 fema… Aust… AUS   2.34e-4 2.34e-4 0.995 2.33e-4 0.995  81.2  81.6     2
#>  4  2003 fema… Aust… AUS   2.07e-4 2.07e-4 0.995 2.06e-4 0.995  80.2  80.6     3
#>  5  2003 fema… Aust… AUS   1.67e-4 1.67e-4 0.995 1.66e-4 0.995  79.2  79.6     4
#>  6  2003 fema… Aust… AUS   8.73e-5 8.73e-5 0.995 8.68e-5 0.995  78.2  78.6     5
#>  7  2003 fema… Aust… AUS   3.89e-5 3.89e-5 0.995 3.87e-5 0.995  77.2  77.6     6
#>  8  2003 fema… Aust… AUS   1.16e-4 1.16e-4 0.995 1.15e-4 0.995  76.2  76.6     7
#>  9  2003 fema… Aust… AUS   1.20e-4 1.20e-4 0.994 1.20e-4 0.994  75.2  75.6     8
#> 10  2003 fema… Aust… AUS   9.04e-5 9.04e-5 0.994 8.99e-5 0.994  74.2  74.6     9
#> # ℹ 91 more rows