Read data from files downloaded from HMD and construct a tsibble object for use in other functions

read_hmd_files reads 1x1 data from files downloaded from the Human Mortality Database (HMD https://www.mortality.org) and constructs a tsibble object suitable for use in other functions. This function uses readHMD to parse the files. At least one file is required. If any file is missing, the corresponding variable will be missing from the output.

read_hmd_files(Deaths = NULL, Exposures = NULL, Population = NULL, Mx = NULL)

Arguments

Deaths

File containing deaths data

Exposures

File containing exposures data

Population

File containing population data

Mx

File containing mortality data

Value

read_hmd_files returns a tsibble object with the following variables:

Year

Numerical vector containing year of observation

Age

Numerical vector containing age group

OpenInterval

Logical vector indicating if the age interval is open

Sex

Character vector taking values "Female", "Male", "Total"

Deaths

Total deaths

Exposure

Numerical vector of exposure-to-risk, equal to population size on 30 June

Population

Numerical vector containing population size on 1 January

Mortality

Numerical vector of mortality rate, equal to Deaths/Exposure

Author

Rob J Hyndman

Examples

if (FALSE) {
read_hmd_files("Deaths_1x1.txt", "Exposures_1x1.txt", "Population.txt", "Mx_1x1.txt")
}