Compute total fertility rate from age-specific fertility rates

Total fertility rate is the expected number of babies per woman in a life-time given the fertility rate at each age of a woman's life.

total_fertility_rate(.data, age, fertility)

Arguments

.data

A tsibble including an age variable and a variable containing fertility rates.

age

Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` or `Age_group` will be used (not case sensitive).

fertility

Variable in `.data` containing fertility rates. If omitted, the variable with name `fx`, `Fertility` or `Rate` will be used (not case sensitive).

Value

A tsibble object with total fertility in column `tfr`.

Examples

# Compute Australian total fertility rates over time
aus_fertility |>
  total_fertility_rate()
#> # A tsibble: 86 x 2 [1Y]
#>     Year   tfr
#>    <int> <dbl>
#>  1  1921 3109.
#>  2  1922 3100.
#>  3  1923 3012.
#>  4  1924 2966.
#>  5  1925 2947.
#>  6  1926 2848.
#>  7  1927 2811.
#>  8  1928 2771.
#>  9  1929 2638.
#> 10  1930 2584.
#> # ℹ 76 more rows